Add Frame Time information
In the UI, it fluctuates quite a lot, but it's still useful to get a
grasp of where the time budget goes.

Note! Time spent in psx::run_frame changes (for example from "Start" to
"Next"), I still don't fully understand why. I guess some timers must be
in place that do this.